refactor(rust): make packages/rust a workspace and extract armonik-transport
`armonik` mixed two concerns: getting a configured, TLS-terminated channel to an ArmoniK endpoint, and
the generated types and services that talk over it. Anyone who wants only the first has to take the
second, which means `protoc` and a `tonic-prost-build` step to compile forty protos they will never call.

`armonik-transport` is that first concern, moved out. `armonik` depends on it and re-exports every type
at the path it already had — `armonik::ClientConfig`, `armonik::client::ConfigError`,
`armonik::client::ConnectionError`, `armonik::client::ReadEnvError` — so nothing downstream has to know
the split happened.

What moved, and only what moved:

- `client/config.rs` → `armonik-transport/src/config.rs`, unchanged but for one `super::ConfigError`
  becoming `ConfigError` now that it is in the defining module.
- The channel building in `Client::with_config`, and the private `Client::https_connector` under it, →
  `connect()` and `https_connector()`. `ConnectionError` moved with them, since it describes exactly
  those failures.
- The environment readers (`read_env`, `read_env_bool`, `ReadEnvError`) and `InsecureCertVerifier` out of
  `utils.rs`. `IntoCollection` and the serde helpers stay: nothing in the transport layer uses them.

`Client::with_config` keeps its `tracing::debug_span!("Client", endpoint)` rather than the span moving
with the code. The span is named after the type, and `armonik-transport` has no `Client` — moving it
would have renamed a span that appears in users' traces, for no reason beyond following the code it
happened to wrap.

Three consequences worth naming rather than leaving to be discovered:

- **`https_connector` is now `pub`, and `#[doc(hidden)]`.** It was private. `armonik`'s `get_nb_request`
  test helper drives a raw HTTP request through the very same TLS stack a channel would use, to read the
  mock server's `/calls.json`; with the connector in another crate that is no longer reachable any other
  way. Hidden because its return type names `hyper-rustls` and `hyper-util` types rather than this
  crate's own.
- **`impl From<ConfigError> for ConnectionError`** is new. `Client::new` built that conversion with
  snafu's generated context selector, which no longer exists on this side of the crate boundary.
- **`hyper` and `hyper-util` become dev-dependencies of `armonik`.** Its two remaining uses of
  `hyper::http::Request` are the same `http` crate reached through `tonic`, which already guarantees the
  matching version; `hyper` itself is now needed only by that test helper. `humantime` and `rustls` and
  `hyper-rustls` leave `armonik` entirely — they went with the code that used them.

The workspace, which the split makes necessary rather than merely tidy: two crates that hand `rustls`
and `tonic` types to each other must not be able to resolve to two different copies of either.
`[workspace.dependencies]` pins the versions once, `[workspace.package]` holds the metadata and the MSRV,
and the lockfile moves to `packages/rust/Cargo.lock` where a workspace's lockfile belongs.

Two things that would otherwise have broken silently:

- `scripts/versions` read the Rust version from `packages/rust/armonik/Cargo.toml`, which no longer
  carries one — `version.workspace = true`. Unfixed, `update-versions` would have skipped Rust without
  saying so and `verify-versions` would have compared one fewer version, so a release could ship a stale
  crate version while looking clean. It now reads `[workspace.package]`, and separately the
  `armonik-transport` dependency version, which has to track the release for `cargo publish` to work at
  all.
- `cargo test`/`build`/`clippy` ran from `packages/rust/armonik`, which covers one member. Both workflows
  now work from `packages/rust` with `--workspace`, or the new crate would have had no CI at all.

Publishing order is now `armonik-transport` first, then `armonik`: a path dependency cannot be published,
so `cargo publish` rewrites it into the version requirement in `armonik`'s manifest, and that version has
to already be in the registry. Written down in the new crate's README.
