+++ packages/rust/armonik-transport/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "armonik-transport"
+description = "Transport layer (configuration, TLS/mTLS) for the ArmoniK Rust client"
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+readme = "README.md"
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+include = ["**/*.rs", "Cargo.toml", "README.md"]
+
+[features]
+serde = ["dep:serde"]
+
+[dependencies]
+# `channel` for `tonic::transport::Endpoint`, which is what `connect` builds; `codegen` for the
+# `http`/`tokio_stream` re-exports `armonik` reaches through this crate.
+tonic = { workspace = true, features = ["channel", "codegen"] }
+snafu.workspace = true
+tracing.workspace = true
+hyper = { workspace = true, features = ["client", "http1", "http2"] }
+hyper-util = { workspace = true, features = ["client", "http1"] }
+hyper-rustls = { workspace = true, features = [
+  "http1",
+  "http2",
+  "native-tokio",
+  "tls12",
+  "logging",
+] }
+rustls = { workspace = true, features = ["ring", "logging", "std", "tls12"] }
+serde = { version = "1.0", features = [
+  "derive",
+], default-features = false, optional = true }
+humantime.workspace = true
