+++ packages/rust/armonik/Cargo.toml
@@ -1,62 +1,51 @@
 [package]
-authors = ["Florian Lemaitre <flemaitre@aneo.fr>"]
+authors.workspace = true
 description = "Rust bindings for the ArmoniK API"
-edition = "2021"
+edition.workspace = true
 include = ["**/*.rs", "Cargo.toml", "LICENSE", "README.md", "protos/**"]
-license = "Apache-2.0"
+license.workspace = true
 readme = "README.md"
 name = "armonik"
-repository = "https://github.com/aneoconsulting/ArmoniK.Api"
-version = "3.29.2-beta-0"
-rust-version = "1.88.0"
+repository.workspace = true
+version.workspace = true
+rust-version.workspace = true
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
 default = ["client"]
-serde = ["dep:serde"]
+serde = ["dep:serde", "armonik-transport?/serde"]
 client = ["_gen-client"]
 server = ["_gen-server"]
 agent = ["_gen-client", "_gen-server"]
 worker = ["_gen-client", "_gen-server"]
-_gen-client = ["tonic/channel", "dep:rustls", "dep:hyper-rustls", "dep:hyper"]
+_gen-client = ["tonic/channel", "dep:armonik-transport"]
 _gen-server = ["tonic/server", "tonic/router", "dep:tokio"]
 
 [dependencies]
-tonic = { version = "0.14", features = ["codegen"], default-features = false }
-tonic-prost = "0.14"
+# TLS, mTLS and the rest of the connection story live in `armonik-transport`; this crate keeps the
+# services and the generated types. By path *and* by version: a path dependency cannot be published, so
+# `cargo publish` rewrites this into the version requirement below, which therefore has to name the
+# version actually being released. `nr update-versions` keeps it in step.
+armonik-transport = { path = "../armonik-transport", version = "3.29.2-beta-0", optional = true }
+tonic = { workspace = true, features = ["codegen"] }
+tonic-prost.workspace = true
 prost = "0.14"
 prost-types = "0.14"
 futures = "0.3"
-snafu = "0.9"
-tracing = "0.1"
-tracing-futures = { version = "0.2", features = ["futures-03"] }
-hyper = { version = "1.10", features = [
-  "client",
-  "http1",
-  "http2",
-], optional = true }
-hyper-util = { version = "0.1", features = ["client", "http1"] }
-hyper-rustls = { version = "0.27", features = [
-  "http1",
-  "http2",
-  "native-tokio",
-  "tls12",
-  "logging",
-], default-features = false, optional = true }
-rustls = { version = "0.23", features = [
-  "ring",
-  "logging",
-  "std",
-  "tls12",
-], default-features = false, optional = true }
-tokio = { version = "1.52", default-features = false, optional = true }
+snafu.workspace = true
+tracing.workspace = true
+tracing-futures = { workspace = true, features = ["futures-03"] }
+tokio = { workspace = true, optional = true }
 serde = { version = "1.0", features = [
   "derive",
 ], default-features = false, optional = true }
-humantime = "2.3.0"
 
 [dev-dependencies]
+# Only the `get_nb_request` test helper needs these: it drives a raw HTTP request through the connector
+# `armonik-transport` builds, to read the mock server's `/calls.json`.
+hyper = { workspace = true, features = ["client", "http1"] }
+hyper-util = { workspace = true, features = ["client", "http1"] }
 eyre = "0.6"
 tracing-subscriber = { version = "0.3", features = ["env-filter", "fmt"] }
 http-body-util = "0.1"
@@ -72,7 +61,7 @@ tokio = { version = "1.48", features = [
 tokio-util = "0.7"
 
 [build-dependencies]
-tonic-prost-build = "0.14"
+tonic-prost-build.workspace = true
 
 [[test]]
 name = "agent"
