+++ packages/rust/armonik/src/client/mod.rs
@@ -1,12 +1,13 @@
 //! ArmoniK clients for all the services
 
-use std::sync::Arc;
+use snafu::Snafu;
 
-use hyper::Uri;
-use hyper_rustls::{ConfigBuilderExt, FixedServerNameResolver, HttpsConnector};
-use hyper_util::client::legacy::connect::HttpConnector;
-use rustls::pki_types::ServerName;
-use snafu::{ResultExt, Snafu};
+// Re-exported at the paths they have always had: the split moved where these are defined, not where
+// they are used from.
+#[cfg(feature = "_gen-client")]
+pub use armonik_transport::{
+    ClientConfig, ClientConfigArgs, ConfigError, ConnectionError, ReadEnvError,
+};
 
 #[cfg(feature = "worker")]
 mod agent;
@@ -14,7 +15,6 @@ mod agent;
 mod applications;
 #[cfg(feature = "client")]
 mod auth;
-mod config;
 #[cfg(feature = "client")]
 mod events;
 #[cfg(feature = "client")]
@@ -34,14 +34,12 @@ mod versions;
 #[cfg(feature = "agent")]
 mod worker;
 
-pub use crate::utils::ReadEnvError;
 #[cfg(feature = "worker")]
 pub use agent::Agent;
 #[cfg(feature = "client")]
 pub use applications::Applications;
 #[cfg(feature = "client")]
 pub use auth::Auth;
-pub use config::{ClientConfig, ClientConfigArgs, ConfigError};
 #[cfg(feature = "client")]
 pub use events::Events;
 #[cfg(feature = "client")]
@@ -71,7 +69,7 @@ pub struct Client<T = tonic::transport::Channel> {
 impl Client<tonic::transport::Channel> {
     /// Create a new client using the configuration from the environment variables
     pub async fn new() -> Result<Self, ConnectionError> {
-        Self::with_config(ClientConfig::from_env().context(ConfigSnafu {})?).await
+        Self::with_config(ClientConfig::from_env()?).await
     }
 
     /// Create a new client with the specified client configuration
@@ -79,124 +77,13 @@ impl Client<tonic::transport::Channel> {
         let endpoint = config.endpoint.to_string();
         tracing_futures::Instrument::instrument(
             async move {
-                let endpoint = config.endpoint.clone();
-                let override_target = config.override_target.clone();
-                let http2_keep_alive_interval = config.http2_keep_alive_interval;
-                let http2_keep_alive_timeout = config.http2_keep_alive_timeout;
-                let http2_keep_alive_while_idle = config.http2_keep_alive_while_idle;
-                let http2_max_header_list_size = config.http2_max_header_list_size;
-                let user_agent = config.user_agent.clone();
-
-                let https = Self::https_connector(config).await?;
-
-                let mut transport_endpoint = tonic::transport::Endpoint::from(endpoint.clone());
-                if let Some(target) = override_target {
-                    transport_endpoint = transport_endpoint.origin(target);
-                }
-
-                if let Some(interval) = http2_keep_alive_interval {
-                    transport_endpoint = transport_endpoint.http2_keep_alive_interval(interval);
-                }
-                if let Some(timeout) = http2_keep_alive_timeout {
-                    transport_endpoint = transport_endpoint.keep_alive_timeout(timeout);
-                }
-                transport_endpoint =
-                    transport_endpoint.keep_alive_while_idle(http2_keep_alive_while_idle);
-                if let Some(max) = http2_max_header_list_size {
-                    transport_endpoint = transport_endpoint.http2_max_header_list_size(max);
-                }
-                if let Some(ua) = user_agent {
-                    transport_endpoint = transport_endpoint
-                        .user_agent(ua)
-                        .expect("HeaderValue is already validated, conversion is infallible");
-                }
-
-                // Build the actual channel from the configuration
-                let channel = transport_endpoint
-                    .connect_with_connector(https)
-                    .await
-                    .context(TransportSnafu { endpoint })?;
-
-                Ok(Self::with_channel(channel))
+                Ok(Self::with_channel(
+                    armonik_transport::connect(config).await?,
+                ))
             },
             tracing::debug_span!("Client", endpoint),
         )
         .await
-    }
-
-    async fn https_connector(
-        config: ClientConfig,
-    ) -> Result<HttpsConnector<HttpConnector>, ConnectionError> {
-        let endpoint = config.endpoint;
-
-        // Get the default crypto provider or fallback to the ring crypto provider
-        let crypto_provider = rustls::crypto::CryptoProvider::get_default()
-            .cloned()
-            .unwrap_or_else(|| Arc::new(rustls::crypto::ring::default_provider()));
-
-        // Configure TLS with sane protocol defaults
-        let tls_config = rustls::ClientConfig::builder_with_provider(crypto_provider)
-            .with_safe_default_protocol_versions()
-            .with_context(|_| TlsSnafu {
-                endpoint: endpoint.clone(),
-            })?;
-
-        // Configure the server verification
-        let tls_config = if config.allow_unsafe_connection {
-            // Do not verify the server
-            tls_config
-                .dangerous()
-                .with_custom_certificate_verifier(Arc::new(crate::utils::InsecureCertVerifier))
-        } else if let Some(cacert) = config.cacert {
-            // Verify that the server certificate is signed with a specific CA cert
-            let mut root_cert_store = rustls::RootCertStore::empty();
-            root_cert_store.add(cacert).with_context(|_| TlsSnafu {
-                endpoint: endpoint.clone(),
-            })?;
-            tls_config.with_root_certificates(root_cert_store)
-        } else {
-            // Verify the server certificate using the system CAs
-            tls_config
-                .with_native_roots()
-                .with_context(|_| IoSnafu {})?
-        };
-
-        // Configure client identity for mTLS
-        let tls_config = if let Some((cert, key)) = config.identity {
-            // Use the the specified client certificate and key for the client authentication
-            tls_config
-                .with_client_auth_cert(vec![cert], key)
-                .with_context(|_| TlsSnafu {
-                    endpoint: endpoint.clone(),
-                })?
-        } else {
-            // No mTLS
-            tls_config.with_no_client_auth()
-        };
-
-        // Configure the connector to use http or https depending on the URI scheme
-        let mut https = hyper_rustls::HttpsConnectorBuilder::new()
-            .with_tls_config(tls_config)
-            .https_or_http();
-
-        if let Some(hostname) = &config.override_target {
-            let server_name = ServerName::try_from(hostname.host().unwrap_or_default())
-                .expect("A valid URI host should be a valid ServerName")
-                .to_owned();
-            https = https.with_server_name_resolver(FixedServerNameResolver::new(server_name));
-        };
-
-        let mut http = HttpConnector::new();
-        http.enforce_http(false); // required for hyper-rustls to switch schemes
-        http.set_nodelay(!config.tcp_nagle_algorithm);
-        http.set_keepalive(config.tcp_keepalive);
-        http.set_keepalive_interval(config.tcp_keepalive_interval);
-        http.set_keepalive_retries(config.tcp_keepalive_retries);
-        if let Some(timeout) = config.connect_timeout {
-            http.set_connect_timeout(Some(timeout));
-        }
-
-        Ok(https.enable_http1().enable_http2().wrap_connector(http))
     }
 
     #[cfg(test)]
@@ -222,7 +109,7 @@ impl Client<tonic::transport::Channel> {
             .body(http_body_util::Empty::<&[u8]>::new())
             .expect("Request");
 
-        let https = Self::https_connector(config)
+        let https = armonik_transport::https_connector(config)
             .await
             .expect("Build connection information");
 
@@ -408,7 +295,7 @@ where
         self.channel.poll_ready(cx)
     }
 
-    fn call(&mut self, request: hyper::http::Request<tonic::body::Body>) -> Self::Future {
+    fn call(&mut self, request: tonic::codegen::http::Request<tonic::body::Body>) -> Self::Future {
         self.channel.call(request)
     }
 }
@@ -431,7 +318,7 @@ where
         self.channel.poll_ready(cx)
     }
 
-    fn call(&mut self, request: hyper::http::Request<tonic::body::Body>) -> Self::Future {
+    fn call(&mut self, request: tonic::codegen::http::Request<tonic::body::Body>) -> Self::Future {
         self.channel.call(request)
     }
 }
@@ -471,45 +358,6 @@ where
     async fn call(self, request: Stream) -> Result<Self::Response, Self::Error> {
         <T as GrpcCallStream<Request, Stream>>::call(self, request).await
     }
-}
-
-#[derive(Debug, Snafu)]
-#[non_exhaustive]
-pub enum ConnectionError {
-    #[snafu(display("Could not read the client config [{location}]"))]
-    #[non_exhaustive]
-    Config {
-        #[snafu(source(from(ConfigError, Box::new)))]
-        source: Box<ConfigError>,
-        #[snafu(implicit)]
-        location: snafu::Location,
-    },
-    #[snafu(display("Could not connect to the remote {endpoint} [{location}]"))]
-    #[non_exhaustive]
-    Transport {
-        endpoint: Uri,
-        #[snafu(source(from(tonic::transport::Error, Box::new)))]
-        source: Box<tonic::transport::Error>,
-        #[snafu(implicit)]
-        location: snafu::Location,
-    },
-    #[snafu(display("Could not establish TLS connection to the remote {endpoint} [{location}]"))]
-    #[non_exhaustive]
-    Tls {
-        endpoint: Uri,
-        #[snafu(source(from(rustls::Error, Box::new)))]
-        source: Box<rustls::Error>,
-        #[snafu(implicit)]
-        location: snafu::Location,
-    },
-    #[snafu(display("Could not read system cert store [{location}]"))]
-    #[non_exhaustive]
-    Io {
-        #[snafu(source(from(std::io::Error, Box::new)))]
-        source: Box<std::io::Error>,
-        #[snafu(implicit)]
-        location: snafu::Location,
-    },
 }
 
 #[derive(Debug, Snafu)]
