+++ packages/rust/armonik/src/utils.rs
@@ -1,5 +1,3 @@
-use snafu::Snafu;
-
 pub(crate) trait IntoCollection<T> {
     fn into_collect(self) -> T;
 }
@@ -13,109 +11,6 @@ where
 {
     fn into_collect(self) -> TY {
         self.into_iter().map(Into::into).collect()
-    }
-}
-
-pub(crate) fn read_env(name: &str) -> Result<String, ReadEnvError> {
-    match std::env::var(name) {
-        Ok(value) => Ok(value),
-        Err(std::env::VarError::NotPresent) => Ok(String::new()),
-        Err(std::env::VarError::NotUnicode(value)) => NotUnicodeSnafu {
-            name: name.to_owned(),
-            value,
-        }
-        .fail(),
-    }
-}
-
-pub(crate) fn read_env_bool(name: &str) -> Result<bool, ReadEnvError> {
-    let value = read_env(name)?;
-    match value.as_ref() {
-        "0" | "false" | "no" | "disable" | "disallow" | "forbid" | "" => Ok(false),
-        "1" | "true" | "yes" | "enable" | "allow" | "authorize" => Ok(true),
-        _ => NotBooleanSnafu {
-            name: name.to_owned(),
-            value,
-        }
-        .fail(),
-    }
-}
-
-#[derive(Debug, Snafu)]
-#[non_exhaustive]
-pub enum ReadEnvError {
-    #[snafu(display(
-        "Environment variable `{name}={value:?}` is not a valid unicode string [{location}]"
-    ))]
-    #[non_exhaustive]
-    NotUnicode {
-        name: String,
-        value: std::ffi::OsString,
-        backtrace: snafu::Backtrace,
-        #[snafu(implicit)]
-        location: snafu::Location,
-    },
-    #[snafu(display("Environment variable `{name}={value}` is not a valid boolean [{location}]"))]
-    #[non_exhaustive]
-    NotBoolean {
-        name: String,
-        value: String,
-        backtrace: snafu::Backtrace,
-        #[snafu(implicit)]
-        location: snafu::Location,
-    },
-}
-
-#[derive(Debug)]
-pub(crate) struct InsecureCertVerifier;
-
-#[cfg(feature = "_gen-client")]
-impl rustls::client::danger::ServerCertVerifier for InsecureCertVerifier {
-    fn verify_server_cert(
-        &self,
-        _end_entity: &rustls::pki_types::CertificateDer<'_>,
-        _intermediates: &[rustls::pki_types::CertificateDer<'_>],
-        _server_name: &rustls::pki_types::ServerName<'_>,
-        _ocsp_response: &[u8],
-        _now: rustls::pki_types::UnixTime,
-    ) -> Result<rustls::client::danger::ServerCertVerified, rustls::Error> {
-        Ok(rustls::client::danger::ServerCertVerified::assertion())
-    }
-
-    fn verify_tls12_signature(
-        &self,
-        _message: &[u8],
-        _cert: &rustls::pki_types::CertificateDer<'_>,
-        _dss: &rustls::DigitallySignedStruct,
-    ) -> Result<rustls::client::danger::HandshakeSignatureValid, rustls::Error> {
-        Ok(rustls::client::danger::HandshakeSignatureValid::assertion())
-    }
-
-    fn verify_tls13_signature(
-        &self,
-        _message: &[u8],
-        _cert: &rustls::pki_types::CertificateDer<'_>,
-        _dss: &rustls::DigitallySignedStruct,
-    ) -> Result<rustls::client::danger::HandshakeSignatureValid, rustls::Error> {
-        Ok(rustls::client::danger::HandshakeSignatureValid::assertion())
-    }
-
-    fn supported_verify_schemes(&self) -> Vec<rustls::SignatureScheme> {
-        vec![
-            rustls::SignatureScheme::RSA_PKCS1_SHA1,
-            rustls::SignatureScheme::ECDSA_SHA1_Legacy,
-            rustls::SignatureScheme::RSA_PKCS1_SHA256,
-            rustls::SignatureScheme::ECDSA_NISTP256_SHA256,
-            rustls::SignatureScheme::RSA_PKCS1_SHA384,
-            rustls::SignatureScheme::ECDSA_NISTP384_SHA384,
-            rustls::SignatureScheme::RSA_PKCS1_SHA512,
-            rustls::SignatureScheme::ECDSA_NISTP521_SHA512,
-            rustls::SignatureScheme::RSA_PSS_SHA256,
-            rustls::SignatureScheme::RSA_PSS_SHA384,
-            rustls::SignatureScheme::RSA_PSS_SHA512,
-            rustls::SignatureScheme::ED25519,
-            rustls::SignatureScheme::ED448,
-        ]
     }
 }
